services/agents/calendar_agent/requirements.txt
python-dotenv==1.0.0
python-dateutil==2.8.2
google-cloud-storage==2.10.0

# Metrics server
prometheus-client==0.19.0
flask==3.0.0
